model/product-list: separate server code
